+++ src/map.h
@@ -4,12 +4,10 @@
 #ifndef STUGE_MAP_H
 #define STUGE_MAP_H
 
-#include <stdint.h>
-
 typedef struct {
     char* name;
     char* buf;
-    size_t size;
+    int size;
     int script_idx;
     bool script_executed;
     tmx_map* tmx_map;
@@ -19,7 +17,7 @@ void* TMX_TextureLoader(const char *path);
 void TMX_FreeTexture(void *ptr);
 
 void LoadTilesets(void);
-void LoadMap(size_t idx);
+void LoadMap(int idx);
 
 void Map_Update(void);
 void Map_Draw(void);
